fix double to float in questions_table

+++ database/migrations/2019_03_18_003140_create_questions_table.php
@@ -16,7 +16,7 @@ class CreateQuestionsTable extends Migration
         Schema::create('questions', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('form_id')->nullable()->default(2);
-            $table->double('rog_num', 4, 2)->nullable();
+            $table->float('rog_num', 8, 2)->nullable();
             $table->char('subpart', 2)->nullable();
             $table->text('body');
             $table->char('response_type', 5)->default('text');
